+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added modal transition systems (MTSs), modal contracts (MCs) and Membership-MCs (MMCs).
 * Added `SubsequentialTransducer` interface and implementations/utilities.
 * Added support for systems of procedural automata (SPAs) as well as related concepts (equivalence, etc.).
+* Added M3C model-checker for verifying µ-calculus and CTL formulae on modal context-free process systems (thanks to @AlnisM).
 
 ### Changed